Fixed text boxes going off screen. Modfied text box sizes
Fixed text boxes being too large. This was caused by a constant text thickness not working well with different sized images. Made thickness dynamic based on the width of the image. This still causes the text to be different sizes (relative to size of image) but it looks better overall. If even more relative text size stability is needed we can try to find a function of thickness that keeps the ratio of the size of the text box generated  to the size of image constant.

If the text box goes off the right side of the image we just shift it over to the left to be back in frame. If the text box goes above the image we just put it under the bounding box instead.